Resolve PR Review Comments
